% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_commands.R
\name{compile_commands}
\alias{compile_commands}
\title{Parse R expressions into parallel CUDA code and compile}
\usage{
compile_commands(expr_ls)
}
\arguments{
\item{expr_ls}{A character string representing the unparsed expression
for which the dimensions are to be identified.}
}
\value{
A commands objects that holds the meta information necessary for
executing the compiled commands from a specified environment.
}
\description{
Parses the list of input R expressions into CUDA code, including dimensional
information executed on CPU prior to GPU device call. Also compiles the 
generated machine code into a pseduo R package that can be called using
run_commands.
}
\examples{
expr_ls <- list(substitute(x <- x + 1))

commands_object <- compile_commands(expr_ls)

run_commands(commands_object, environment())
}
