% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_start_end_lines.R
\name{construct_flags}
\alias{construct_flags}
\title{Create the character flags to search for in the compiled code}
\usage{
construct_flags(flag_str = c("Kernel", "Iter.lens", "Expr.lens", "Int.evals"))
}
\arguments{
\item{flag_str}{A character string representing the type of start and end
flags that are to be constructed.  Must be one of "Kernel", "Iter.lens",
"Expr.lens", or "Int.evals".}
}
\value{
list of start and end flag strings for the input flag type
}
\description{
Construct start and end flag strings to be matched against lines 
in machine updated kernel.cu file
}
\examples{
construct_flags("Kernel")
}
