% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_start_end_lines.R
\name{find_start_end_lines}
\alias{find_start_end_lines}
\title{Identify the region of text to replace in lines of compiled code}
\usage{
find_start_end_lines(
  text_lines,
  flag_str = c("Kernel", "Iter.lens", "Expr.lens", "Int.evals")
)
}
\arguments{
\item{text_lines}{A character vector that represents the lines of compiled
code that will be used to search for the start and end indices of the
provided type}

\item{flag_str}{A character string representing the type of start and end
flags that are to be constructed.  Must be one of "Kernel", "Iter.lens",
"Expr.lens", or "Int.evals".}
}
\value{
Vector of start and end indices for the specified flag_str
}
\description{
Identifies the start and end line indices of a specified character string
that represents the type of machine generated code that will be updated
}
\examples{
find_start_end_lines("Kernel")
}
