% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_additional_lines.R
\name{get_additional_lines}
\alias{get_additional_lines}
\title{Retrieve intermediate evaluation lines from parsed arguments}
\usage{
get_additional_lines(parsed_args)
}
\arguments{
\item{parsed_args}{A list of character vectors that represent parsed lines
of code which include the current arguments and any intermediate evaluation
lines needed before those arguments are used.}
}
\value{
A character vector of lines of code for all intermediate evaluations
needed before the current arguments are used.
}
\description{
Takes arguments representing the lines of machine generated code
that may or may not have additional lines for each argument.
Iterate through the arguments and return a single vector of all
intermediate evaluation lines that must be executed prior to 
the final expression that uses these arguments
}
\examples{
get_additional_lines(parsed_args)
}
