% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intermediate_evaluation.R
\name{get_intermediate_evaluation}
\alias{get_intermediate_evaluation}
\title{Writes code to evaluate and save nested matrix argument}
\usage{
get_intermediate_evaluation(arg, var_names)
}
\arguments{
\item{arg}{A character string representing the argument that may or may not
be a nested matrix function  If this arg is not a simple reference to a 
global Rvar, the function proceeds to generate code to execute the nested
matrix function and save the result in an intermediate evaluation Rvar.}

\item{var_names}{A character vector that represents the the named R variables
included in these commands.}
}
\value{
Character vector representing the lines of code needed to evaluate
the nested matrix function and save the results in the newly created 
intermediate Rvar that will then be referenced by the parent expression.
}
\description{
Writes compiled code necessary to execute a nested matrix function, then save
the results of that expression into an intermediate evaluation Rvar that will
be retrieved as the argument of the parent expression calling this function.
}
\examples{
get_intermediate_evaluation(arg, var_names)
}
