% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_args.R
\name{identify_args}
\alias{identify_args}
\title{Identify all arguments of a racket like R expression}
\usage{
identify_args(expr_chars)
}
\arguments{
\item{expr_chars}{A character string representing a function call R 
expression in racket like format for which the function string itself
has already been removed from consideration, leaving only the arguments
to be identified.}
}
\value{
A character vector representing the distinct arguments of 
the current expression.
}
\description{
Repeatedly iterate through a string representing a racket like R expression,
specifically, an expression with the general form (f ...) with any number of 
arguments and identify the all arguments of the expression.
}
\examples{
identify_args(expr_chars)
}
