% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent_lines.R
\name{indent_lines}
\alias{indent_lines}
\title{Indent input lines}
\usage{
indent_lines(lines, depth = 1)
}
\arguments{
\item{lines}{A character vector representing the lines of text to indent}

\item{depth}{An integer representing the number of indentations to apply}
}
\value{
A character vector representing the input lines with the requested
indentation applied
}
\description{
Use a pre specified indentation to indent the lines input by the specified
depth of indentation.
}
\examples{
indent_lines(lines, depth)
}
