% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_interpreter.R
\name{init_interpreter}
\alias{init_interpreter}
\title{Initialize interpreter variables for specified dimensional information}
\usage{
init_interpreter(var_names, env_key = c("loop", "assign", "int_eval"))
}
\arguments{
\item{var_names}{A character vector that represents the the named R variables
included in these commands.}

\item{env_key}{A character string identifying the environment tracked 
globally by the interpreter that stores dimensional information for iteration 
loops, assignment expressions, or intermediate evaluation Rvars for nested 
matrix function calls.}
}
\description{
Initializes globally tracked variables used for parsing dimensional 
information in the compiled commands.The variables initialized will be
used to write the dimensional information to the initialize_iter_lens(),
initialize_expr_lens, or initialize_int_evals() functions in kernel.cu 
depending on which environment is specified.
}
\examples{
init_interpreter(var_names, env_to_init)
}
