% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter.R
\name{interpreter}
\alias{interpreter}
\title{Parse input expressions and write code to the kernel}
\usage{
interpreter(expr_ls, var_names)
}
\arguments{
\item{expr_ls}{A list of R expressions to be parsed, interpreted, and 
then written to the kernel.}

\item{var_names}{A character vector that represents the the named R variables
included in these commands.}
}
\description{
Top level function used to parse and interpret the expressions that need
to be translated to CUDA code, as well as write the machine generated code
to the kernel.
}
\examples{
interpreter(expr_ls, var_names)
}
