% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_start_end_lines.R
\name{match_flags}
\alias{match_flags}
\title{Matches compiled code flags against an input}
\usage{
match_flags(input_str, start_flag_str, end_flag_str)
}
\arguments{
\item{input_str}{A character string representing the line of text that 
will be searched for a match of either the start or end flag}

\item{start_flag_str}{A character string representing the start flag}

\item{end_flag_str}{A character string representing the end flag}
}
\value{
Integer representing a flag that indicates whether to start, end,
or not relevant.
}
\description{
Match the input start and end flags against the input character
vector which is expected to be a line of code in kernel.cu. Return integer
that represents either the start index, end index, or other, which is ignored.
}
\examples{
match_flags("Kernel")
}
