% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr_dim.R
\name{parse_expr_dim}
\alias{parse_expr_dim}
\title{Parse dimensional information of expression}
\usage{
parse_expr_dim(expr_chars, var_names)
}
\arguments{
\item{expr_chars}{A character string representing the unparsed expression
for which the dimensions are to be identified.}

\item{var_names}{A character vector that represents the the named R variables
included in these commands.}
}
\value{
A named list with 3 elements, the element of 'len' representing 
the length of the expression, the element of 'rdim' representing the number
of rows in the expression, and the element of 'cdim' representing the number
of columns in the expression.
}
\description{
Recursive function which identifies the dimensional information, specifically
length, number of rows, and number of columns, of the input R expression.
}
\examples{
parse_expr_dim(expr_chars, var_names)
}
