% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racket_string.R
\name{racket_string}
\alias{racket_string}
\title{Translate an R expression into a racket like string}
\usage{
racket_string(input_expr)
}
\arguments{
\item{input_expr}{An R expression that will be translated into a racket 
like character string.}
}
\value{
A character string that represents the translated expression in the
form of (fun ...) with any arbitrary level of nested arguments also 
translated to the same format.
}
\description{
Recursive function which translates an R expression with up to any arbitrary 
level of nested arguments into a racket like string of the form '(fun ...)'.
}
\examples{
racket_string(input_expr)
}
