% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_commands.R
\name{run_commands}
\alias{run_commands}
\title{Run compiled commands in the specified environment}
\usage{
run_commands(compiled_commands, eval_env)
}
\arguments{
\item{compiled_commands}{A commands object generated by compile_commands that
can be executed in any specified environment, assuming the variables listed
in the commands object are defined in that environment.}
}
\description{
Executes the commands object in the specified environment by calling 
the peudo package which is indexed by the in the commands object key.
}
\examples{
expr_ls <- list(substitute(x <- x + 1))

commands_object <- compile_commands(expr_ls)

run_commands(commands_object, environment())
}
