% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dim_info.R
\name{save_dim_info}
\alias{save_dim_info}
\title{Saves dimensional information of iteration loop}
\usage{
save_dim_info(arg, env_to_update)
}
\arguments{
\item{arg}{A character string representing the expression that must be parsed
for dimensional information.}

\item{env_to_update}{A environment that stores the necessary dimensional 
information for assignment expressions, for loops, or intermediate matrix
evaluations.}
}
\description{
Saves the necessary dimensional information relating to an expression. The 
expression's dimension may be related to updating values of global variables, 
controlling the number of iterations in a for loop, or an intermediate
evaluation of a nested matrix argument.
}
\examples{
save_dim_info(arg, env_to_update)
}
