% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_expr.R
\name{write_expr}
\alias{write_expr}
\title{Parse and interpret an R expression into CUDA code}
\usage{
write_expr(expr, var_names)
}
\arguments{
\item{expr}{An R expression to be parsed and interpreted into CUDA code}

\item{var_names}{A character vector that represents the the named R variables
included in these commands.}
}
\value{
A character vector representing the lines of code necessary
to execute the input expression in CUDA code
}
\description{
High level function used to parse and interpret a single R expression and
then generate CUDA code to execute that R expression on the GPU
}
\examples{
write_expr(expr, var_names)
}
