% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_for_loop.R
\name{write_for_loop}
\alias{write_for_loop}
\title{Writes code to execute a for loop}
\usage{
write_for_loop(args, var_names, depth)
}
\arguments{
\item{args}{A character vector representing the arguments provided in the
parsed R expression. The first is the iteration variable that will be 
updated at each iteration.  The second is the iterable object that will be
used to update the iteration variable at each step.  The third is the 
expression to be executed at each iteration.}

\item{var_names}{A character vector that represents the the named R variables
included in these compiled command.}

\item{depth}{A integer representing the depth of this loop, as nested loops
are allowed.  This will be 0 on a top level call to write_for_loop, and 
increase by 1 for each nested call.}
}
\value{
Character vector representing the lines of code needed to execute
the for loop expression on the kernel.
}
\description{
Writes compiled code necessary to execute a for loop, first by updating the
iteration variable identified as the first argument in the R expression to
be equal to the ith value of the second argument on iteration i, then 
executes the 3rd argument at each iteration after that update has completed.
}
\examples{
write_for_loop(args, var_names, depth)
}
